# Add WorkProcessorOperatorAdapter and use it for TopNOperator