--- a/presto-main/src/test/java/io/prestosql/operator/TestTopNOperator.java
+++ b/presto-main/src/test/java/io/prestosql/operator/TestTopNOperator.java
@@ -24,7 +24,6 @@ import io.prestosql.sql.planner.plan.PlanNodeId;
 import io.prestosql.testing.MaterializedResult;
 import org.testng.annotations.AfterMethod;
 import org.testng.annotations.BeforeMethod;
-import org.testng.annotations.DataProvider;
 import org.testng.annotations.Test;
 
 import java.util.List;
@@ -74,14 +73,8 @@ public class TestTopNOperator
         scheduledExecutor.shutdownNow();
     }
 
-    @DataProvider(name = "useWorkProcessorOperator")
-    public static Object[][] useWorkProcessorOperator()
-    {
-        return new Object[][] {{true}, {false}};
-    }
-
-    @Test(dataProvider = "useWorkProcessorOperator")
-    public void testSingleFieldKey(boolean useWorkProcessorOperator)
+    @Test
+    public void testSingleFieldKey()
     {
         List<Page> input = rowPagesBuilder(BIGINT, DOUBLE)
                 .row(1L, 0.1)
@@ -97,7 +90,6 @@ public class TestTopNOperator
                 .build();
 
         OperatorFactory operatorFactory = topNOperatorFactory(
-                useWorkProcessorOperator,
                 ImmutableList.of(BIGINT, DOUBLE),
                 2,
                 ImmutableList.of(0),
@@ -111,8 +103,8 @@ public class TestTopNOperator
         assertOperatorEquals(operatorFactory, driverContext, input, expected);
     }
 
-    @Test(dataProvider = "useWorkProcessorOperator")
-    public void testMultiFieldKey(boolean useWorkProcessorOperator)
+    @Test
+    public void testMultiFieldKey()
     {
         List<Page> input = rowPagesBuilder(VARCHAR, BIGINT)
                 .row("a", 1L)
@@ -127,7 +119,6 @@ public class TestTopNOperator
                 .build();
 
         OperatorFactory operatorFactory = topNOperatorFactory(
-                useWorkProcessorOperator,
                 ImmutableList.of(VARCHAR, BIGINT),
                 3,
                 ImmutableList.of(0, 1),
@@ -142,8 +133,8 @@ public class TestTopNOperator
         assertOperatorEquals(operatorFactory, driverContext, input, expected);
     }
 
-    @Test(dataProvider = "useWorkProcessorOperator")
-    public void testReverseOrder(boolean useWorkProcessorOperator)
+    @Test
+    public void testReverseOrder()
     {
         List<Page> input = rowPagesBuilder(BIGINT, DOUBLE)
                 .row(1L, 0.1)
@@ -159,7 +150,6 @@ public class TestTopNOperator
                 .build();
 
         OperatorFactory operatorFactory = topNOperatorFactory(
-                useWorkProcessorOperator,
                 ImmutableList.of(BIGINT, DOUBLE),
                 2,
                 ImmutableList.of(0),
@@ -173,12 +163,11 @@ public class TestTopNOperator
         assertOperatorEquals(operatorFactory, driverContext, input, expected);
     }
 
-    @Test(dataProvider = "useWorkProcessorOperator")
-    public void testLimitZero(boolean useWorkProcessorOperator)
+    @Test
+    public void testLimitZero()
             throws Exception
     {
         OperatorFactory factory = topNOperatorFactory(
-                useWorkProcessorOperator,
                 ImmutableList.of(BIGINT),
                 0,
                 ImmutableList.of(0),
@@ -192,8 +181,8 @@ public class TestTopNOperator
         }
     }
 
-    @Test(dataProvider = "useWorkProcessorOperator")
-    public void testExceedMemoryLimit(boolean useWorkProcessorOperator)
+    @Test
+    public void testExceedMemoryLimit()
             throws Exception
     {
         List<Page> input = rowPagesBuilder(BIGINT)
@@ -205,7 +194,6 @@ public class TestTopNOperator
                 .addDriverContext();
 
         OperatorFactory operatorFactory = topNOperatorFactory(
-                useWorkProcessorOperator,
                 ImmutableList.of(BIGINT),
                 100,
                 ImmutableList.of(0),
@@ -220,26 +208,17 @@ public class TestTopNOperator
     }
 
     private OperatorFactory topNOperatorFactory(
-            boolean workProcessorOperator,
             List<? extends Type> types,
             int n,
             List<Integer> sortChannels,
             List<SortOrder> sortOrders)
     {
-        OperatorFactory factory = new TopNOperatorFactory(
+        return new TopNOperatorFactory(
                 0,
                 new PlanNodeId("test"),
                 types,
                 n,
                 sortChannels,
                 sortOrders);
-        if (workProcessorOperator) {
-            factory = new TestWorkProcessorOperator.TestWorkProcessorOperatorFactory(
-                    99,
-                    new PlanNodeId("test"),
-                    (WorkProcessorOperatorFactory) factory);
-        }
-
-        return factory;
     }
 }