… presto-main/src/main/java/io/prestosql/operator/TopNWorkProcessorOperator.java
… presto-main/src/test/java/io/prestosql/operator/TestWorkProcessorOperator.java