some benchmarks:
```
Before best
Benchmark                   (positionsPerPage)  (topN)   Mode  Cnt   Score   Error  Units
BenchmarkTopNOperator.topN                  32       1  thrpt   60  23,024 ± 0,292  ops/s
BenchmarkTopNOperator.topN                  32     100  thrpt   60  22,101 ± 0,249  ops/s
BenchmarkTopNOperator.topN                  32   10000  thrpt   60   6,634 ± 0,178  ops/s
BenchmarkTopNOperator.topN                1024       1  thrpt   60  32,031 ± 0,227  ops/s
BenchmarkTopNOperator.topN                1024     100  thrpt   60  29,742 ± 0,180  ops/s
BenchmarkTopNOperator.topN                1024   10000  thrpt   60   9,917 ± 0,091  ops/s

After 1
Benchmark                   (positionsPerPage)  (topN)   Mode  Cnt   Score   Error  Units
BenchmarkTopNOperator.topN                  32       1  thrpt   60  21,778 ± 0,293  ops/s
BenchmarkTopNOperator.topN                  32     100  thrpt   60  20,971 ± 0,463  ops/s
BenchmarkTopNOperator.topN                  32   10000  thrpt   60   6,601 ± 0,193  ops/s
BenchmarkTopNOperator.topN                1024       1  thrpt   60  31,952 ± 0,199  ops/s
BenchmarkTopNOperator.topN                1024     100  thrpt   60  29,581 ± 0,389  ops/s
BenchmarkTopNOperator.topN                1024   10000  thrpt   60   9,756 ± 0,089  ops/s

After 2
Benchmark                   (positionsPerPage)  (topN)   Mode  Cnt   Score   Error  Units
BenchmarkTopNOperator.topN                  32       1  thrpt   60  22,921 ± 0,333  ops/s
BenchmarkTopNOperator.topN                  32     100  thrpt   60  21,170 ± 0,249  ops/s
BenchmarkTopNOperator.topN                  32   10000  thrpt   60   6,637 ± 0,186  ops/s
BenchmarkTopNOperator.topN                1024       1  thrpt   60  31,663 ± 0,267  ops/s
BenchmarkTopNOperator.topN                1024     100  thrpt   60  30,293 ± 0,217  ops/s
BenchmarkTopNOperator.topN                1024   10000  thrpt   60   9,694 ± 0,243  ops/s
```